[RUI-251]: Get rid of pro utils
